requirements.txt
flask
pinecone
requests
python-dotenv